requirements.txt
fastapi
uvicorn
requests
python-multipart
python-docx  # 新增：生成Word文档